Fix hub selection highlight bug in BotListPanel
Added selectedHub to state becasue the active hub info needed to be
separated from detailsBoxItem which holds the info for the most recent
click in the BotListPanel. Otherwise, if the user clicked (in the panel)
the hub, then a bot, and then the hub again, React would register it as
the hub being clicked on, not off.