Limit panel item selection to one item at a time
Now when a bot is clicked in the panel, the state property:
'detailsBoxItem' will be updated with the bot info. Previously, this
property was only used when a user clicked on the hub.

This property is passed into the BotListPanel component and used in the
function get_selected_panel_item_id to determine which panel
(hub or bot) is selected so it can be highlighted accordingly.