Remove 'editing' property from RunInterface
The 'editing' property was only changed by the commented-out edit toggle
on the mission panel. 'isActive' is a more robust implementation of the
'editing' property since it accounts for all the ways a run can be
started and toggles the edit mode on or off accordingly. It can be used
if the edit run toggle is re-implemented in the mission panel.