Fix bug for executing runs that delete first, then begin the run
Bug stemmed from the lag in the API call to stop the runs when they are
deleted. It is not necessary to stop the runs, if a new run is going to
be immediately triggered such as 'go to green rally'. However, it is
necessary when the operator plans to delete runs and then manually add
news runs. If the bots are not stopped, edit mode will not be entered.