Feature/add-toggle-for-edit-run-1 (#483)
* Prevent operator from editing a run once it starts

* Signal editability through the colors of the waypoints and lines
* Send warnings to operator when they attempt to edit an active run

* Change edit + select colors + coordinate colors with bots

* Remove 'editing' property from RunInterface

The 'editing' property was only changed by the commented-out edit toggle
on the mission panel. 'isActive' is a more robust implementation of the
'editing' property since it accounts for all the ways a run can be
started and toggles the edit mode on or off accordingly. It can be used
if the edit run toggle is re-implemented in the mission panel.

* Prevent edit mode from breaking JDV by passing necessary arguments

* Add toggle to MissionPanel

* Edit mode now turns back on when a mission is paused or completed
* Fixed duplicate mission goals copy bug

* Make toggle switch to off when run begins

Now, setEditRunMode is only used in one file, RunItem.tsx

* Fix bug for executing runs that delete first, then begin the run

Bug stemmed from the lag in the API call to stop the runs when they are
deleted. It is not necessary to stop the runs, if a new run is going to
be immediately triggered such as 'go to green rally'. However, it is
necessary when the operator plans to delete runs and then manually add
news runs. If the bots are not stopped, edit mode will not be entered.

* Merge 1.y without conflicts

* Fix merge conflict from 1.y pull

* Fix CommandControl.tsx merge conflicts

* Resolve merge conflicts from Task/refactor command control 2

* Remove border-rad on details boxes to make them flush with the botsList

* Keep edit mode off after run stops

* Prevent a mission active run from being deleted

Also cleaned up the disable button function for the Details panel.
Instead of returning a custom object, they now return a boolean.

* Modiy enabled states + fix delete-all-runs bug + add line tool check

* Disable clear run btn in RunItem + update delete all runs warning msg

---------

Co-authored-by: ferrom-jaia <109534520+ferrom-jaia@users.noreply.github.com>